requirements.txt
streamlit==1.45.0
python-dotenv==1.1.0
langchain==0.3.25
langchain-openai==0.2.14
langchain-community==0.3.23
faiss-cpu==1.11.0
crewai[tools]==0.44.0
